requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
telnyx==2.0.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
openai>=1.6.1,<2.0.0
anthropic==0.7.7
langchain==0.0.350
langchain-openai==0.0.2
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx>=0.25.2
celery==5.3.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Pluto Memory and Learning Dependencies
redis>=4.5.0
numpy>=1.21.0
apscheduler>=3.10.0
pytz>=2023.3  # Timezone support for digests
freezegun>=1.2.0  # For time-based testing

# Additional dependencies for enhanced features
slack-sdk>=3.21.0  # Slack API integration
